# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "根据用户名和密码登录",
                "parameters": [
                    {
                        "description": "登录表单",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "100": {
                        "description": "Continue",
                        "schema": {
                            "$ref": "#/definitions/auth.ResponseResult"
                        }
                    },
                    "104": {
                        "schema": {
                            "$ref": "#/definitions/auth.ResponseResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "auth.ResponseResult": {
            "$ref": "#/definitions/handler.ResponseResult"
        },
        "handler.ResponseResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 100
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "msg": {
                    "description": "提示",
                    "type": "string",
                    "example": "密码错误"
                }
            }
        }
    }
}